database/pyproject.toml
[tool.poetry]
name = "generic_microservice_database"
version = "0.1.0"
description = "The database submodule for the generic KEVEN microservice, responsible for data persistence and storage operations."
authors = ["Craig Holland <craig_holland@hotmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "database"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
protobuf = "^4.25.0"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"

# Internal Dependencies (referencing other submodules)
generic_microservice_domain = "0.1.0"
alembic = "^1.15.1"
psycopg2 = "^2.9.10"
sqla-wrapper = "3.0.0"
sqlalchemy = "1.4.42"
sqlalchemy-utils = "0.41.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.1"
mypy = "^1.5.1"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
start = "services.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "local"
url = "http://localhost:9090"
priority = "supplemental"



